Add files via upload
Hospital management system is basically a system which helps efficiently store the record of the patients admitted to the hospital along with the employee’s detail.
In this system, the entities are patient, employees, rooms for patients, and patient’s record. When a patient is admitted for a certain treatment in a hospital, the system efficiently stores the admitted room number for the patient, assigned doctors for treatment, assigned nurses, other employees involved along with the span of treatment, the patient record like when to meet the doctor again after the discharge. Furthermore, the system stores record of medicine that the patients have to buy until he recovers totally. In addition, the system will store the record- name, address, phone number, of the employees as well, In fact, the certain doctors who are specialized in certain treatments. Hence the system is useful to store all kind of record prevailed in a hospital